Add local authentication strategy and user registration
- Implement local authentication strategy using passport-local
- Add bcrypt for password hashing
- Create user registration endpoint
- Update user schema to support local authentication
- Modify GitHub authentication to be optional
- Add login endpoint for local authentication
- Update user service with new methods for local auth
